# docs-website/static/docs/api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for the traduora REST API\n\nOfficial website: https://traduora.com\nAdditional documentation: https://docs.traduora.com\nSource code: https://github.com/traduora/traduora",
    "version": "0.14.0",
    "title": "Traduora API"
  },
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "password",
      "authorizationUrl": "/api/v1/auth/token",
      "tokenUrl": "/api/v1/auth/token"
    }
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/terms": {
      "get": {
        "summary": "List a project's terms",
        "parameters": [
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ListProjectTermsResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "post": {
        "summary": "Add a new project term",
        "parameters": [
          {
            "name": "AddTermRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddTermRequest"
            }
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ProjectTermResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Plan limit reached"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
    "/api/v1/projects/{projectId}/terms/{termId}": {
      "patch": {
        "summary": "Update a project's term",
        "parameters": [
          {
            "name": "UpdateTermRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTermRequest"
            }
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/ProjectTermResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Remove a project's term",
        "description": "Removes a project's term and all related translations",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
    "/api/v1/projects/{projectId}/labels/{labelId}/terms/{termId}": {
      "post": {
        "summary": "Label a project term",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Unlabel a project term",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Label removed"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
  },
  "definitions": {
...
    "AccessDatesDTO": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string"
        },
        "modified": {
          "type": "string"
        }
      },
      "required": ["created", "modified"]
    },
...
    "ProjectTermDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "labels": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "date": {
          "$ref": "#/definitions/AccessDatesDTO"
        }
      },
      "required": ["id", "value", "labels", "date"]
    },
    "ListProjectTermsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProjectTermDTO"
          }
        }
      },
      "required": ["data"]
    },
    "AddTermRequest": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "required": ["value"]
    },
    "ProjectTermResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ProjectTermDTO"
        }
      },
      "required": ["data"]
    },
    "UpdateTermRequest": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "required": ["value"]
    },
...
  }
}